Break up 'Deconvoluter' class
