+++ src/python/spectrseqtools/preprocessing.py
@@ -1,10 +1,14 @@
 import numpy as np
 import polars as pl
+import tqdm
 import yaml
 
-from spectrseqtools.common import set_output_path
-from spectrseqtools.deconvolution import Deconvoluter
+from spectrseqtools.common import set_output_path, initialize_raw_file_iterator
+from spectrseqtools.deconvolution import DeconvolutionParameters, DeisotopedPeakList
 from spectrseqtools.singleton_identification import identify_singletons
+
+
+MIN_MS1_CHARGE_STATE = 3
 
 
 class Preprocessor:
@@ -33,9 +37,7 @@ class Preprocessor:
         print("RAW file found. Preprocessing raw data...")
 
         # Deconvolute raw data from file
-        fragments = Deconvoluter(params=self.deconvolution_params).deconvolute(
-            file_path=str(self.input_path),
-        )
+        fragments = self.deconvolute(file_path=str(self.input_path))
 
         # Update meta parameters (if needed)
         meta_params = self.meta_params
@@ -66,6 +68,48 @@ class Preprocessor:
         singletons.write_csv(f"{self.output_prefix}.singletons.tsv", separator="\t")
 
         print("Preprocessing completed!\n")
+
+    def deconvolute(self, file_path: str) -> pl.DataFrame:
+        """
+        Deconvolute/deisotope peaks in MS2 scans from ThermoFisher RAW file.
+
+        Parameters
+        ----------
+        file_path : str
+            Path of RAW file from ThermoFisher.
+
+        Returns
+        -------
+        polars.DataFrame
+            Dataframe containing fragment monoisotopic masses and intensities.
+
+        """
+        # Load deconvolution parameter based on parameter dict
+        params = DeconvolutionParameters(self.deconvolution_params)
+
+        # Initialize iterator for RAW file
+        raw_file_read = initialize_raw_file_iterator(file_path=file_path)
+
+        peak_list = DeisotopedPeakList.default()
+        for _ in tqdm.tqdm(range(len(raw_file_read) - 1), desc="Deisotoping MS2 scans"):
+            # Select next scan
+            scan = next(raw_file_read)
+
+            # Skip scan if it is no MS2 scan
+            if scan.ms_level != 2:
+                continue
+
+            # Skip scan if the precursor charge is lower than MIN_MS1_CHARGE_STATE
+            if (
+                not isinstance(scan.precursor_information.charge, int)
+                or scan.precursor_information.charge < MIN_MS1_CHARGE_STATE
+            ):
+                continue
+
+            # Deconvolute scan to get list of deisotoped peaks
+            peak_list += DeisotopedPeakList.from_scan(scan=scan, params=params)
+
+        return peak_list.to_fragments()
 
     def select_intact_mass(self, fragments: pl.DataFrame) -> float:
         """
